requirements.txt
altair==5.5.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.34
botocore==1.38.34
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.4
cycler==0.11.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==2.2.6
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pickleshare==0.7.5
pillow==11.1.0
pip==25.1
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
sip==6.10.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.15.0
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.45.1
yarg==0.1.9